common/image: Allow initrd_high & fdt_high to have special case no_reloc
The documented way of disabling relocation of the initrd or FDT is
setting initrd_high or fdt_high to 0xffffffff.
However, that is actually wrong on 64-bit platforms, where
0xffffffffffffffff needs to be used. This is somewhat painful:

 - Some platforms (e.g. Raspberry Pi) have a common file setting fdt_high
   or initrd_high to 0xffffffff that is included in both 32-bit builds
   (RPi 1, 2) and 64-bit builds (RPi 3). Those are currently not doing
   the right thing in 64-bit builds, and it wouldn't be nice to #ifdef
   around that.
 - Typing the 16 consecutive f's in 0xffffffffffffffff is not very
   pleasant.

To fix these, make the initrd_high and fdt_high handling accept a special
value "no_reloc" that does the right thing independent of the bitness.
The old values keep working for compatibility.

Signed-off-by: Tuomas Tynkkynen <tuomas@tuxera.com>